reorder tests
