+++ package.json
@@ -26,6 +26,7 @@
     "runtime-tools": "^0.3.0",
     "runtimeify": "^0.2.1",
     "runtimejs": "^0.1.9",
+    "split": "^1.0.0",
     "tap-finished": "0.0.1",
     "tape": "^4.0.0"
   },
